refactor: move common types into releasy-core crate
