+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "1"
-members = ["releasy-graph", "releasy-emit"]
+members = ["releasy-graph", "releasy-emit", "releasy-core"]
