+++ releasy-core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "releasy-core"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+reqwest = "0.11.20"
+serde = "1.0.188"
+serde_json = "1.0.107"
+thiserror = "1.0.48"
