+++ releasy-core/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod error;
+pub mod event;
+pub mod repo;
+
+pub mod default {
+    pub const DEFAULT_MANIFEST_FILE_NAME: &str = "repo-plan.toml";
+}
