+++ releasy-emit/Cargo.toml
@@ -7,6 +7,7 @@ exclude = ["/tests"]
 [dependencies]
 anyhow = "1.0.75"
 clap = { version = "4.4.2", features = ["derive"] }
+releasy-core = { path = "../releasy-core" }
 releasy-graph = { path = "../releasy-graph/" }
 reqwest = "0.11.20"
 serde = { version = "1.0.188", features = ["derive"] }
