+++ releasy-graph/Cargo.toml
@@ -7,6 +7,7 @@ description = "A utility library for describing cross-repo relations in the form
 
 [dependencies]
 petgraph = "0.6.4"
+releasy-core = { path = "../releasy-core/" }
 serde = { version = "1.0.188", features = ["derive"] }
 serde_ignored = "0.1.9"
 thiserror = "1.0.48"
